Move offset logic to utils and also adjust connection line (TODO: Remove dependency on our ids)
